// src/interfaces/User.ts
export interface User{
    address: {
        city: string
        geo: {
            lat: string
            lng: string
        }
        street: string
        suite: string
        zipcode: string
    }
    company: {
        name: string
        catchPhrase: string
        bs: string
    }
    email: string
    id: number
    name: string
    phone: string
    username: string
    website: string
}